bilbo contribution!! Mummi contribution!! veriverse contribution!! LCTV contribution!! epryka, zzzel, markus, bananastriker, reddevil and gosu contribution!!
Thank you Carlos!!

Thank you Mummi!!

Carol contribution!!

at bilbo at veriverse!!
i7 6700 16GB RAM SSD OCZ 120GB SSD 240GB HD 2TB Windows 10 Pro

+++ appseed/aura/primitive/datetime/datetime.cpp
@@ -572,7 +572,7 @@ CLASS_DECL_AURA DWORD get_tick_count()
 CLASS_DECL_AURA DWORD get_first_tick()
 {
 
-   return (DWORD)(get_first_nano() / (uint64_t)(1000 * 1000));
+   return (DWORD)((uint64_t)get_first_nano() / (uint64_t)(1000 * 1000));
 
 }
